+++ source/LibUISharp/src/LibUISharp/Drawing/FontStretch.cs
@@ -4,7 +4,7 @@ namespace LibUISharp.Drawing
     /// <summary>
     /// Specifies how a font will be stretched compared to the normal aspect ratio of that font.
     /// </summary>
-    public enum FontStretch : uint
+    public enum FontStretch : long
     {
         /// <summary>
         /// Specifies the ultra-condensed font stretch.
