+++ source/LibUISharp/src/LibUISharp/Drawing/FontStyle.cs
@@ -4,7 +4,7 @@ namespace LibUISharp.Drawing
     /// <summary>
     /// Specifies the style of a font as normal, italic or oblique.
     /// </summary>
-    public enum FontStyle : uint
+    public enum FontStyle : long
     {
         /// <summary>
         /// Specifies the normal font style.
