+++ source/LibUISharp/src/LibUISharp/RelativeAlignment.cs
@@ -4,7 +4,7 @@ namespace LibUISharp
     /// <summary>
     /// Specifies how contents are positioned in relation to other content.
     /// </summary>
-    public enum RelativeAlignment : uint
+    public enum RelativeAlignment : long
     {
         /// <summary>
         /// The contents are positioned before the other content.
